backend/requirements.txt
Django == 3.0.7
djangorestframework == 3.11.0
flake8 == 3.8.2
django-cors-headers == 3.3.0
django-filter == 2.3.0
drf-schema-adapter == 2.0.7
